restructure shared/api 폴더 분리
