+++ pages/app/plans/[id].tsx
@@ -1,5 +1,5 @@
 import { PlanDetail } from '@/pages/app/plans/[id]';
-import { fetchPlanDetailSSR } from '@/shared/api/ssr/plans';
+import { fetchPlanDetailSSR } from './ssr/plans';
 import { QUERY_KEY } from '@/shared/constants/queryKey';
 import {
   dehydrate,
