+++ src/shared/hooks/useJoinMeetingMutation.tsx
@@ -1,4 +1,7 @@
-import { joinMeeting, leaveMeeting } from '@/shared/api/meeting';
+import {
+  joinMeeting,
+  leaveMeeting,
+} from '@/features/update-meeting/api/meeting';
 import { QUERY_KEY } from '@/shared/constants/queryKey';
 import { MeetingDetailResponse } from '@/shared/types/api/meeting';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
